add users list with fetch example and corres. minimal test
