+++ examples/cookbook/app/network-requests/index.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import UsersHome from './UsersHome';
+
+export default function Example() {
+  return <UsersHome />;
+}
